requirements.txt
# Для полноценной разработки телеграм-ботов рекомендуется:
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.6
redis==5.0.1
sentry-sdk==1.38.0
prometheus-client==0.19.0
cryptography==41.0.7
schedule==1.2.0
flask==2.3.3
